refactor: Move tests to right place
